requirements.txt
customtkinter
tkinterdnd2
pillow